Initial TypeScript Port
- Stop ignoring package-lock.json (nor recommended)
- Bump all dependencies
- Configure TypeScript Compiler
- Fix several unhandled potential errors
- Add many comments
- Migrate to TypeScript
